Use YAML file to manage inputs and refactor
This is an initial refactor that utilises a yaml file but keeps the same
methodology as before.  We provide only a simulation path on the command line.
 In a config directory we also provide a yaml file which allows the user to specify
Grid_T (3D) files, (T, U, V), resampled grid_T (2D), and restart file.
Ths commit history is compressed - the code logic works - but subsequent commits
improve on the data handling.

* Fix bug due to time issue and clean docstrings
* Add DINO-setup.yaml config
* Add config read logic
* Pull metric output and loader to modules
* Move metric output to metrics_io.py
* Move file I/O logic to loader.py
* Add module names and docstrings
* Fix loader bug (e3t_0) on restart file
* Rename variable_aliases to standardise_inputs.py
* Rewrite comment at bottom of README.md
* Correct standardise imports and rename fn